tools/requirements.txt
pandas
pillow
feedparser
